refactor: moved music data to data folder
